Database + CRUD

+++ backend/src/main/java/com/dia/backend/data/repository/MovieRepository.java
@@ -3,4 +3,4 @@ package com.dia.backend.data.repository;
 import com.dia.backend.domain.model.Movie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MovieRepository extends JpaRepository<Movie, Long> {}
+public interface MovieRepository extends JpaRepository<Movie, Integer> {}
